+++ template/backend/src/Ambev.DeveloperEvaluation.Application/Sales/UpdateSale/UpdateSaleCommand.cs
@@ -24,6 +24,6 @@ namespace Ambev.DeveloperEvaluation.Application.Sales.UpdateSale
         public Guid Id { get; set; }
         public Guid ProductId { get; set; }
         public int Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
+        public long UnitPrice { get; set; }
     }
 }
