+++ template/backend/src/Ambev.DeveloperEvaluation.WebApi/Features/Sales/UpdateSale/UpdateSaleRequest.cs
@@ -13,6 +13,6 @@
         public Guid Id { get; set; }
         public Guid ProductId { get; set; }
         public int Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
+        public long UnitPrice { get; set; }
     }
 }
